requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
google-cloud-speech==2.27.0
google-cloud-dialogflow-cx==1.35.0
google-cloud-texttospeech==2.17.2
